refactor: move .ts components from components/ to vue/
Move KoAjaxDialog, KoDialog, and KoStudySearch from vite/src/components/
to vite/src/vue/ to consolidate all Vue-related non-SFC files alongside
the .vue components. Update all import paths accordingly.
