+++ src/vite/src/vue/KoAjaxDialog.ts
