+++ src/vite/src/vue/KoStudySearch.ts
